+++ kibot/__main__.py
@@ -329,6 +329,11 @@ def detect_kicad():
     GS.ki8 = GS.kicad_version_major >= 8
     GS.ki9 = GS.kicad_version_major >= 9
     GS.ki10 = GS.kicad_version_major >= 10
+    if GS.ki10:
+        GS.VIATYPE_THROUGH = pcbnew.VIATYPE_THROUGH
+        GS.VIATYPE_BLIND = pcbnew.VIATYPE_BLIND
+        GS.VIATYPE_BURIED = pcbnew.VIATYPE_BURIED
+        GS.VIATYPE_MICROVIA = pcbnew.VIATYPE_MICROVIA
     GS.footprint_gr_type = 'MGRAPHIC' if not GS.ki8 else 'PCB_SHAPE'
     GS.board_gr_type = 'DRAWSEGMENT' if GS.ki5 else 'PCB_SHAPE'
     GS.footprint_update_local_coords = GS.dummy1 if GS.ki8 else GS.footprint_update_local_coords_ki7