+++ CHANGELOG.md
@@ -17,11 +17,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - KiCad 9 embedded files: populate the cache if needed (See #890)
 - preflights:
   - Mechanism to merge preflights from imports (`merge_pre`) (#900)
+- Global options:
+  - `allow_blind_vias` and `allow_buried_vias`: for KiCad 10 to replace the
+    unified `allow_blind_buried_vias`
 - 3D models: option to download the latest 3D models (`kicad_3d_url_version`)
   (See #899)
 - BoM: descriptive error when using `kicad` format and no options found.
 - DRC/ERC: Configurable logo (#889)
 - iBoM: `mark_when_checked` option (#881)
+- Report: `buried_vias_count` for KiCad 10
 
 ### Fixed
 - 3D models: download models from the repo tag corresponding to the currently